Update asciidoctor-0-1-2-released.adoc
I changed 'issues' to 'resolves issues' because just 'issues' seemed to communicate that there were open problems with the features. Also, since they stood alone on their own lines, they seemed to dominate the page.

I moved the ordering of the improvements so that the substitutions and masquerading were right above the corresponding paragraph.

Added link to issue track for milestone

There are two comments in the text for you to address.