Add additional case requiring unconstrained quotes (#741)
Where the User Manual gives us the conditions under which we need to use unconstrained (rather than constrained)  quotes (Sec. 12.3), it lists a couple of marks which, only when they appear directly before the opening quote, require it:

- Is there a colon or semi-colon directly before the starting formatting mark?

In my experimentation, it seems that a closing curly bracket causes the same necessity. 

This is the case, for example, when you have an attribute reference immediately before the starting formatting mark and you are using the normal substitution ordering of quote-substitution before attribute-ref-substitution.