add information about Working Copy
Working Copy is a Git client for iPhone and iPad and as such somewhat of a niche tool. 

I'm the developer of this Git client and not really sure it should be mentioned on this page, but it on the other hand it does support AsciiDoc rendering and syntax highlighting.