LICENSE for website content
The asciidoctor.org website specifies in its footer that the content on the site is CC BY SA 3.0, and that would seem to be the license that all the files in this repo are under. 

For example, if I am grabbing a template to use as-is or modify for my project, I want to properly attribute that template (presuming I can merge in from a CC license, which I can in my personal case).

It seems that you intend the entirety of this repo to be CC BY SA 3.0.

So I am proposing this LICENSE file to live in the repo with the soruce files to make it clear what the content here is licensed under. This also makes attribution (BY term) easier to follow, because one can link directly to the file in the git repo rather than the rendered HTML page on asciidoctor.org.