Add Codeberg + GitLab to the 'Where' section
GitHub is not the only place where AsciiDoc is supported

Codeberg supports it, with the same level of accuracy as GitHub on rendering (though `include::[]` directives aren't important and the links 404 at this time) where block titles look a bit weird and admonitions are in an ugly table. GitLab is supported as well--and that support is more robust (with `include::[]`) and the styling generally nicer to top it off. This merge request looks to highlight this fact and have people consider using an open-source or open-core code forge over a proprietary, closed source one.

`Codeberg or GitHub or GitLab` is ordered alphabetically.